[1.6.1] `check_batch` raised ArgumentError on Ruby 4.0+ (#84)
