+++ Gemfile_2_7.lock
@@ -1,7 +1,7 @@
 PATH
   remote: .
   specs:
-    docscribe (1.6.0)
+    docscribe (1.6.1)
       parser (>= 3.3)
       prism (~> 1.8)
