+++ lib/docscribe/version.rb
@@ -1,5 +1,5 @@
 # frozen_string_literal: true
 
 module Docscribe
-  VERSION = '1.6.0'
+  VERSION = '1.6.1'
 end
