--- sig/lib/docscribe/server.rbs
@@ -1,117 +1,0 @@
-class Dir
-  def self?.tmpdir: () -> String
-end
-
-module Parser
-  class SyntaxError < StandardError
-    def line: () -> Integer?
-    def diagnostic: () -> untyped
-  end
-end
-
-module Timeout
-  def self?.timeout: ((Integer | Float) sec) { () -> untyped } -> untyped
-
-  class Error < StandardError
-  end
-end
-
-module Docscribe
-  module Server
-    SOCKET_DIR: String
-    IDLE_TIMEOUT: Integer
-    ENV_FILES: Array[String]
-
-    def self?.running?: (?String? config_path) -> bool
-    def self?.read_pid: (?String? config_path) -> Integer?
-    def self?.pid_path: (?String? config_path) -> String
-    def self?.socket_path: (?String? config_path) -> String
-    def self?.config_hash: (String config_path) -> String
-    def self?.env_hash: () -> String
-    def self?.check_platform_support!: () -> void
-    def self?.ensure_running!: (?config_path: String?, ?daemonize: bool, ?timeout: Integer) -> void
-    def self?.wait_for_ready: (?config_path: String?, ?timeout: Integer, ?raise_on_timeout: bool) -> bool
-    def self?.start_daemon_process: (config_path: String?, daemonize: bool) -> void
-    def self?.handle_stale_socket?: (String? config_path) -> bool
-    def self?.process_alive?: (Integer pid) -> bool
-    def self?.clean_socket_files: (String? config_path) -> void
-
-    module Protocol
-      def self?.build_request: (String method, ?Hash[Symbol, untyped] params) -> Hash[Symbol, untyped]
-      def self?.parse_response: (String line) -> Hash[String, untyped]?
-      def self?.serialize: (Hash[untyped, untyped] hash) -> String
-    end
-
-    class Client
-      @socket_path: String
-
-      def initialize: (?String? socket_path, ?config_path: String?) -> void
-      def check: (file: String, ?strategy: Symbol, **untyped) -> Hash[String, untyped]?
-      def fix: (file: String, ?strategy: Symbol, **untyped) -> Hash[String, untyped]?
-
-      def shutdown: () -> Hash[String, untyped]?
-      def ping: () -> Hash[String, untyped]?
-
-      private
-
-      def request: (String method, **untyped) -> Hash[String, untyped]?
-      def connect: [T] { (UNIXSocket) -> T } -> T?
-    end
-
-    class Daemon
-      @socket_path: String
-      @idle_timeout: Integer
-      @last_request_time: Numeric
-      @running: bool
-      @server: UNIXServer?
-      @config: Docscribe::Config?
-      @config_path: String?
-      @core_rbs_provider: untyped
-      @file_cache: Docscribe::LRUCache
-      @effective_config: Docscribe::Config?
-      @started_at: Time
-      @applied_overrides: Hash[String, untyped]?
-
-      ERROR_CODES: Hash[Symbol, Integer]
-
-      def initialize: (?socket_path: String?, ?idle_timeout: Integer, ?config_path: String?) -> void
-      def start: () -> void
-
-      private
-
-      def load_dependencies: () -> void
-      def setup_socket: () -> void
-      def write_pid: () -> void
-      def listen_loop: () -> void
-      def check_idle_timeout: () -> void
-      def accept_client: () -> void
-      def handle_client: (UNIXSocket) -> void
-      def handle_request: (UNIXSocket, Hash[String, untyped]) -> void
-      def handle_check: (UNIXSocket, (String | Integer), Hash[String, untyped]) -> void
-      def handle_fix: (UNIXSocket, (String | Integer), Hash[String, untyped]) -> void
-      def handle_check_batch: (untyped, untyped, untyped) -> void
-      def run_rewrite: (String, Symbol) -> Hash[String, untyped]
-      def apply_cli_overrides: (Hash[String, untyped]?) -> void
-      def build_effective_config: (Hash[String, untyped]) -> void
-      def reset_effective_config: () -> void
-      def reset_effective_config_internal: () -> void
-      def rewrite_file: (String, Symbol) -> [ String, Hash[Symbol, untyped] ]
-      def rewrite_and_cache: (String, Symbol, untyped, untyped, untyped) -> [ String, Hash[Symbol, untyped] ]
-      def process_file_in_batch: (String, Symbol, ?(Integer | Float)?) -> Hash[String, untyped]
-      def handle_shutdown: (UNIXSocket, (String | Integer)) -> void
-      def handle_ping: (UNIXSocket, (String | Integer)) -> void
-      def send_result: (UNIXSocket, (String | Integer), Hash[String, untyped]) -> void
-      def send_error: (UNIXSocket, (String | Integer | nil), Integer, String, ?untyped?) -> void
-      def classify_error: (Exception, ?String?, ?Hash[String, untyped]) -> [ Integer, String, untyped? ]
-      def send_syntax_error: (UNIXSocket, (String | Integer), untyped, String) -> void
-      def handle_request_error: (UNIXSocket, (String | Integer), untyped, String) -> void
-      def syntax_error?: (Exception) -> bool
-      def timeout_error?: (Exception) -> bool
-      def classify_gem_error: (untyped) -> [ Integer, String, untyped ]
-      def classify_syntax_err: (untyped, Hash[String, untyped]) -> [ Integer, String, untyped ]
-      def classify_timeout_err: (untyped, Hash[String, untyped]) -> [ Integer, String, untyped ]
-      def classify_internal_err: (untyped) -> [ Integer, String, untyped ]
-      def cleanup: () -> void
-    end
-  end
-end
