+++ sig/lib/docscribe/server/base.rbs
@@ -1,0 +1,27 @@
+# frozen_string_literal: true
+
+class Dir
+  def self?.tmpdir: () -> String
+end
+
+module Docscribe
+  module Server
+    SOCKET_DIR: String
+    IDLE_TIMEOUT: Integer
+    ENV_FILES: Array[String]
+
+    def self?.running?: (?String? config_path) -> bool
+    def self?.read_pid: (?String? config_path) -> Integer?
+    def self?.pid_path: (?String? config_path) -> String
+    def self?.socket_path: (?String? config_path) -> String
+    def self?.config_hash: (String config_path) -> String
+    def self?.env_hash: () -> String
+    def self?.check_platform_support!: () -> void
+    def self?.ensure_running!: (?config_path: String?, ?daemonize: bool, ?timeout: Integer) -> void
+    def self?.wait_for_ready: (?config_path: String?, ?timeout: Integer, ?raise_on_timeout: bool) -> bool
+    def self?.start_daemon_process: (config_path: String?, daemonize: bool) -> void
+    def self?.handle_stale_socket?: (String? config_path) -> bool
+    def self?.process_alive?: (Integer pid) -> bool
+    def self?.clean_socket_files: (String? config_path) -> void
+  end
+end
