+++ sig/lib/docscribe/server/daemon.rbs
@@ -1,0 +1,87 @@
+# frozen_string_literal: true
+
+module Parser
+  module Source
+    class Map
+      def line: () -> Integer?
+    end
+
+    class Diagnostic
+      def location: () -> Parser::Source::Map
+    end
+  end
+
+  class SyntaxError < StandardError
+    def line: () -> Integer?
+    def diagnostic: () -> Parser::Source::Diagnostic
+  end
+end
+
+module Timeout
+  def self?.timeout: [T] (Integer | Float sec) { () -> T } -> T
+
+  class Error < StandardError
+  end
+end
+
+module Docscribe
+  module Server
+    class Daemon
+      @socket_path: String
+      @idle_timeout: Integer
+      @last_request_time: Numeric
+      @running: bool
+      @server: UNIXServer?
+      @config: Docscribe::Config?
+      @config_path: String?
+      @core_rbs_provider: (Docscribe::Types::RBS::Provider | nil)
+      @file_cache: Docscribe::LRUCache
+      @effective_config: (Docscribe::Config | nil)
+      @cache_mutex: Mutex
+      @config_mutex: Mutex
+      @started_at: Time
+      @applied_overrides: (Hash[String, Object] | nil)
+
+      ERROR_CODES: Hash[Symbol, Integer]
+
+      def initialize: (?socket_path: String?, ?idle_timeout: Integer, ?config_path: String?) -> void
+      def start: () -> void
+
+      private
+
+      def load_dependencies: () -> void
+      def setup_socket: () -> void
+      def write_pid: () -> void
+      def listen_loop: () -> void
+      def check_idle_timeout: () -> void
+      def accept_client: () -> void
+      def handle_client: (UNIXSocket) -> void
+      def handle_request: (UNIXSocket, Hash[String, untyped]) -> void
+      def handle_check: (UNIXSocket, (String | Integer), Hash[String, untyped]) -> void
+      def handle_fix: (UNIXSocket, (String | Integer), Hash[String, untyped]) -> void
+      def handle_check_batch: (UNIXSocket, (String | Integer), Hash[String, untyped]) -> void
+      def process_file_in_batch: (String, Symbol, ?(Integer | Float)?) -> Hash[String, (String | Array[Hash[Symbol, untyped]])]
+      def run_rewrite: (String, Symbol) -> Hash[String, (String | Array[Hash[Symbol, untyped]])]
+      def apply_cli_overrides: (Hash[String, Object]?) -> void
+      def build_effective_config: (Hash[String, Object]) -> void
+      def reset_effective_config: () -> void
+      def reset_effective_config_internal: () -> void
+      def rewrite_file: (String, Symbol) -> [String, { output: String, changes: Array[Hash[Symbol, untyped]] }]
+      def rewrite_and_cache: (String, Symbol, Docscribe::Config, Array[(String | Symbol)], Time) -> [String, { output: String, changes: Array[Hash[Symbol, untyped]] }]
+      def handle_shutdown: (UNIXSocket, (String | Integer)) -> void
+      def handle_ping: (UNIXSocket, (String | Integer)) -> void
+      def send_result: (UNIXSocket, (String | Integer), Hash[String, Object]) -> void
+      def send_error: (UNIXSocket, (String | Integer | nil), Integer, String, ?untyped?) -> void
+      def classify_error: (Exception, ?String?, ?Hash[String, Object]) -> [Integer, String, (Hash[Symbol, (Object | nil)] | nil)]
+      def syntax_error?: (Exception) -> bool
+      def timeout_error?: (Exception) -> bool
+      def classify_gem_error: (untyped) -> [Integer, String, Hash[Symbol, (String | nil)]]
+      def classify_syntax_err: (untyped, Hash[String, Object]) -> [Integer, String, Hash[Symbol, Object]]
+      def classify_timeout_err: (Exception, Hash[String, Object]) -> [Integer, String, Hash[Symbol, Object]]
+      def classify_internal_err: (Exception) -> [Integer, String, Hash[Symbol, Object]]
+      def handle_request_error: (UNIXSocket, (String | Integer), untyped, String) -> void
+      def send_syntax_error: (UNIXSocket, (String | Integer), untyped, String) -> void
+      def cleanup: () -> void
+    end
+  end
+end
