+++ spec/spec_helper.rb
@@ -28,6 +28,10 @@ RSpec.configure do |config|
   config.include RbsHelper
   config.include StreamHelper
   config.include SarifHelper
+  config.include SuppressErrorHelper
+  config.include CleanFileHelper
+  config.include ServerWireHelper
+  config.include DaemonHelper
   config.example_status_persistence_file_path = '.rspec_status'
   config.disable_monkey_patching!
   config.expect_with(:rspec) { |c| c.syntax = :expect }
